requirements.txt
fastapi
uvicorn[standard]
firebase-admin
python-jose
python-multipart
bcrypt
